Adopt stdint.h types for portability
- Use uintptr_t + PRIxPTR in hex_string_from(void*) and
  hex_string_from(void(*)()) instead of reinterpret_cast to
  unsigned long long, which is not guaranteed to hold a pointer
- Change get_time_in_millis return type from unsigned long to
  uint_least64_t to avoid narrowing from 64-bit platform APIs
  (GetTickCount64, gettimeofday) on Win64 where unsigned long is 32-bit
- Store time fields in Result and JUnit output as uint_least64_t;
  route time values through string_from() at print call sites to
  avoid overload ambiguity on 32-bit platforms where uint_least64_t
  differs from size_t

+++ src/test/Result.cpp
@@ -5,6 +5,8 @@
 
 #include "mutiny/time.hpp"
 
+#include <stdint.h>
+
 namespace mu {
 namespace tiny {
 namespace test {
@@ -17,20 +19,20 @@ Result::Result(Output& p)
 void Result::current_group_started(Shell* test)
 {
   output_.print_current_group_started(*test);
-  current_group_time_started_ = static_cast<size_t>(get_time_in_millis());
+  current_group_time_started_ = get_time_in_millis();
 }
 
 void Result::current_group_ended(Shell* /*test*/)
 {
   current_group_total_execution_time_ =
-      static_cast<size_t>(get_time_in_millis()) - current_group_time_started_;
+      get_time_in_millis() - current_group_time_started_;
   output_.print_current_group_ended(*this);
 }
 
 void Result::current_test_started(Shell* test)
 {
   output_.print_current_test_started(*test);
-  current_test_time_started_ = static_cast<size_t>(get_time_in_millis());
+  current_test_time_started_ = get_time_in_millis();
 }
 
 void Result::print(const char* text)
@@ -46,7 +48,7 @@ void Result::print_very_verbose(const char* text)
 void Result::current_test_ended(Shell* /*test*/)
 {
   current_test_total_execution_time_ =
-      static_cast<size_t>(get_time_in_millis()) - current_test_time_started_;
+      get_time_in_millis() - current_test_time_started_;
   output_.print_current_test_ended(*this);
 }
 
@@ -88,33 +90,33 @@ void Result::count_ignored()
 
 void Result::tests_started()
 {
-  time_started_ = static_cast<size_t>(get_time_in_millis());
+  time_started_ = get_time_in_millis();
   output_.print_tests_started();
 }
 
 void Result::tests_ended()
 {
-  auto time_ended = static_cast<size_t>(get_time_in_millis());
+  auto time_ended = get_time_in_millis();
   total_execution_time_ = time_ended - time_started_;
   output_.print_tests_ended(*this);
 }
 
-size_t Result::get_total_execution_time() const
+uint_least64_t Result::get_total_execution_time() const
 {
   return total_execution_time_;
 }
 
-void Result::set_total_execution_time(size_t ex_time)
+void Result::set_total_execution_time(uint_least64_t ex_time)
 {
   total_execution_time_ = ex_time;
 }
 
-size_t Result::get_current_test_total_execution_time() const
+uint_least64_t Result::get_current_test_total_execution_time() const
 {
   return current_test_total_execution_time_;
 }
 
-size_t Result::get_current_group_total_execution_time() const
+uint_least64_t Result::get_current_group_total_execution_time() const
 {
   return current_group_total_execution_time_;
 }
